+++ llm/rtsp/rtsp.h
@@ -13,8 +13,8 @@ extern "C" {
 //rtsp_init.c
 rtsp_packet_t* generate_rtsp_packets(int count);
 
-size_t parse_rtsp_msg(const uint8_t *buf, size_t buf_len,
-                           rtsp_packet_t *out_packets, size_t max_count);
+size_t parse_rtsp_msg(const uint8_t *buf, u32 buf_len,
+                           rtsp_packet_t *out_packets, u32 max_count);
 void print_rtsp_packets(const rtsp_packet_t *packets, size_t count) ;
 
 //rtsp_mutators.c
