+++ llm/sip/sip.h
@@ -9,7 +9,7 @@
 
 sip_packet_t* generate_sip_packets(int count);
 
-size_t parse_sip_msg(const uint8_t *buf, size_t buf_len, sip_packet_t *out_packets, size_t max_count);
+size_t parse_sip_msg(const uint8_t *buf, u32 buf_len, sip_packet_t *out_packets, u32 max_count);
 
 int reassemble_sip_msgs(const sip_packet_t *packets, u32 num_packets, u8 *output_buf, u32 *out_len);
