+++ llm/smtp/smtp_parser.c
@@ -111,8 +111,8 @@ static const char* next_token(const char *b, const char *e,
 }
 
 
-size_t parse_smtp_msg(const uint8_t *buf, size_t buf_len,
-                      smtp_packet_t *out_packets, size_t max_count)
+size_t parse_smtp_msg(const uint8_t *buf, u32 buf_len,
+                      smtp_packet_t *out_packets, u32 max_count)
 {
     if (!buf || !out_packets || max_count == 0) return 0;
